On branch master

+++ setupTermuxArch.sh
@@ -7,8 +7,18 @@
 
 printf '\033]2;  Thank you for using `setupTermuxArch.sh` 📲 \007'"\n 🕛 \033[36;1m< 🕛 This setup script will attempt to set Arch Linux up in your Termux environment.  When successfully completed, you will be enjoying the bash prompt in Arch Linux in Termux on your smartphone or tablet.  If you do not see 🕐 one o'clock below after updating Termux and installing the required components for Arch Linux in Termux installation is completed, run this script again. You might want to check your Internet connection too.  \033[37;1m\n\n	1)	Updating Termux.  \n\n	2)	Installing the script components for Arch Linux in Termux installation.  \n\n	3)	Activating termux-wake-lock.  \n\n		__This will take some time.__  \033[0m\n\n"
 
+if [ ! -f $PREFIX/bin/proot ]; then
 apt-get -qq update && apt-get -qq upgrade --yes
 apt-get -qq install bsdtar proot wget --yes 
+elif [ ! -f $PREFIX/bin/bsdtar ]; then
+apt-get -qq update && apt-get -qq upgrade --yes
+apt-get -qq install bsdtar proot wget --yes 
+elif [ ! -f $PREFIX/bin/wget ]; then
+apt-get -qq update && apt-get -qq upgrade --yes
+apt-get -qq install bsdtar proot wget --yes 
+else
+	printf
+fi
 
 depend ()
 {
